Correções em update e delete de produtos e avaliações

+++ backend/avalie-aqui/src/reviews/reviews.service.ts
@@ -110,16 +110,60 @@ export class ReviewsService {
     throw new HttpException('Avaliação inválida.', HttpStatus.BAD_REQUEST);
   }
 
+  /**
+   * Atualiza os dados de uma avaliação
+   * @param updateReviewDto
+   * @returns
+   */
   async update(id: number, updateReviewDto: Partial<UpdateReviewDto>) {
-    return await this.prisma.review.update({
-      where: {id},
-      data: updateReviewDto
-    });
+    try {
+      const review = await this.prisma.review.findUnique({
+        where: {
+          id: id,
+        },
+      });
+      if (review) {
+        // Atualiza os dados da avaliação
+        await this.prisma.review.update({
+          where: { id: id },
+          data: updateReviewDto,
+        });
+        return { message: 'Avaliação atualizada com sucesso.' };
+      }
+    } catch (error) {
+      throw new HttpException(
+        'Falha ao atualizar dados da avaliação. Tente novamente mais tarde.',
+        HttpStatus.INTERNAL_SERVER_ERROR,
+      );
+    }
+    throw new HttpException('Avaliação inválida.', HttpStatus.BAD_REQUEST);
   }
 
+  /**
+   * Deleta uma avaliação
+   * @param id
+   * @returns
+   */
   async remove(id: number) {
-    return await this.prisma.review.delete({
-      where: {id: id},
-    });
+    try {
+      const review = await this.prisma.review.findUnique({
+        where: {
+          id: id,
+        },
+      });
+      if (review) {
+        // Deleta a avaliação
+        await this.prisma.review.delete({
+          where: { id: id },
+        });
+        return { message: 'Avaliação excluída com sucesso.' };
+      }
+    } catch (error) {
+      throw new HttpException(
+        'Falha ao excluir avaliação. Tente novamente mais tarde.',
+        HttpStatus.INTERNAL_SERVER_ERROR,
+      );
+    }
+    throw new HttpException('Avaliação inválida.', HttpStatus.BAD_REQUEST);
   }
 }
